Add files via upload
Fake news is a growing concern in today’s digital age, as it can cause social unrest and breakage of bonds among people. The classification of fake news is a topic of ongoing research. In this context, the problem statement is to develop an automated system that can accurately detect and classify fake news using machine learning algorithms. The proposed solution involves preprocessing the news data, converting text into vectors, and training a machine learning model to predict the authenticity of news articles. The goal is to create a reliable and efficient system that can help prevent the spread of fake news.